feat(modularization): Phase 5 native GraphQL/gRPC execute in clawql-api
Move native protocol executors into clawql-api and drop ExecuteEnvironment.
executeClawqlOperation and makeExecuteLive now take an optional LoadSpecFn
only; all execute paths run inside the gateway package.
