+++ packages/clawql-api/src/execute/index.ts
@@ -1,5 +1,7 @@
 export * from "./execute-core.js";
 export * from "./execute-live.js";
 export * from "./field-projection.js";
+export * from "./native-grpc.js";
+export * from "./native-graphql.js";
 export * from "./rest-operation.js";
 export * from "./types.js";
